# Conflicts:
#	app/src/main/java/com/example/androidcourseshpp/data/source/network/dto/auth/SignInDTO.kt
#	app/src/main/java/com/example/androidcourseshpp/data/source/network/dto/contacts/AddContactDTO.kt
#	app/src/main/java/com/example/androidcourseshpp/data/source/network/dto/contacts/GetUserContactsDTO.kt
#	app/src/main/java/com/example/androidcourseshpp/data/source/network/model/UserModel.kt
#	app/src/main/java/com/example/androidcourseshpp/data/source/network/model/auth/SignInModel.kt
#	app/src/main/java/com/example/androidcourseshpp/data/source/network/model/auth/SignUpModel.kt
#	app/src/main/java/com/example/androidcourseshpp/data/source/network/model/contacts/GetUserContactsResponseModel.kt
#	app/src/main/java/com/example/androidcourseshpp/data/source/network/repository/UserRepositoryImpl.kt
#	app/src/main/java/com/example/androidcourseshpp/ui/screens/main/addcontacts/AddContactsFragment.kt
#	app/src/main/java/com/example/androidcourseshpp/ui/utils/Extensions.kt

+++ app/src/main/java/com/example/androidcourseshpp/data/source/network/dto/contacts/AddContactDTO.kt
@@ -1,8 +1,7 @@
 package com.example.androidcourseshpp.data.source.network.dto.contacts
-
 import com.google.gson.annotations.SerializedName
 
 data class AddContactRequestDTO(
     @SerializedName("contactId")
-    val contactId: Int
+    val contactId: Long
 )
