+++ app/src/main/java/com/example/androidcourseshpp/domain/repository/UserLocalRepository.kt
@@ -1,8 +1,8 @@
 package com.example.androidcourseshpp.domain.repository
 
 interface UserLocalRepository {
-    fun saveUserServerId(userServerId: Int)
-    fun getUserServerId(): Int
+    fun saveUserServerId(userServerId: Long)
+    fun getUserServerId(): Long
     fun clearUserServerId()
     fun saveUserAvatarUrl(avatar: String)
     fun getUserAvatarUrl(): String
